app_mobile/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
Werkzeug==3.1.3
PyJWT==2.10.1
numpy==2.0.2
opencv-python==4.11.0.86
tensorflow==2.18.0
keras==3.7.0
minio==7.2.15
ultralytics==8.3.69
Pillow==11.0.0 